requirements.txt
discord.py
google-generativeai
py-polymarket
pyyaml
python-dotenv
